% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uri.R
\name{browse_dmr}
\alias{browse_dmr}
\title{Launch URI for browsing if possible}
\usage{
browse_dmr(what = c("mapserver", "info")[1])
}
\arguments{
\item{what}{string, 'mapserver' or 'info'}
}
\value{
integer flag, 0 means success
}
\description{
Launch URI for browsing if possible
}
