% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{has_menssp}
\alias{has_menssp}
\title{Return the default path, possible stored as an R option "MENSSP_PATH"}
\usage{
has_menssp(x, root = options("MENSSP_PATH")[[1]])
}
\arguments{
\item{x}{subdirectory (ala "current" or "nssp" or "P90")}

\item{root}{the path to the various NSSP datasets}
}
\value{
named logical, TRUE if the path exists
}
\description{
The function attempts to retrieve this from option("MENSSP_PATH") which can be
set at the begging of the R session or placed in the users .Rprofile
}
