% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{menssp_path}
\alias{menssp_path}
\title{Retrieve the MENSSP path}
\usage{
menssp_path(...)
}
\arguments{
\item{...}{further arguments for \code{has_nsspme}}
}
\value{
the known MENSSP path
}
\description{
Retrieve the MENSSP path
}
