% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nssp-package.R
\docType{package}
\name{nssp-package}
\alias{nssp-package}
\alias{nssp}
\title{nssp package}
\description{
Access, query and download from RESTful Maine DMR ESRI Map 
   Service.
}
\details{
Access Maine DMR NSSP data files (local or online).  Datasets are stored in 
 geojson format and loaded easily into simple features (sf).  These points
 and polygons can be manipulated/filtered using dplyr and sf packages.
}
