% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{PA_ID}
\alias{PA_ID}
\title{Retrieve sets of PA_NUMBERs by regional name.}
\usage{
PA_ID(region = "Casco Bay")
}
\arguments{
\item{region}{string, defaults to 'Casco Bay'}
}
\value{
a string vector, possibly empty
}
\description{
Retrieve sets of PA_NUMBERs by regional name.
}
\seealso{
\url{http://www.maine.gov/dmr/shellfish-sanitation-management/closures/pollution.html}
}
