% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_archived}
\alias{read_archived}
\title{Read locally stored archived NSSP data}
\usage{
read_archived(year = "2013", path = menssp_path("nssp"))
}
\arguments{
\item{year}{numeric or string year to read}

\item{path}{local path to files}
}
\value{
sf of MULTIPOLYGON or NULL
}
\description{
Read locally stored archived NSSP data
}
