% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_current}
\alias{read_current}
\title{Read the current NSSP (either from disk or online)}
\usage{
read_current(what = c("local", "live")[1],
  local_path = "/Users/ben/Dropbox/shellfish/data/dmr/current", ...)
}
\arguments{
\item{what}{string, either 'local' or 'live'}

\item{local_path}{string, if local look here for the NSSP data}

\item{...}{further arguments for fetch_current if required}
}
\value{
sf
}
\description{
Read the current NSSP (either from disk or online)
}
