% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_nssp}
\alias{read_nssp}
\title{Read locally stored NSSP annual data}
\usage{
read_nssp(year = 2016, path = "/Users/ben/Dropbox/shellfish/data/dmr/nssp")
}
\arguments{
\item{year}{numeric or string year to read}

\item{path}{local path to files}
}
\value{
sf of Points
}
\description{
Read locally stored NSSP annual data
}
