% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_P90}
\alias{read_P90}
\title{Read locally stored P90 data}
\usage{
read_P90(year = 2016, path = "/Users/ben/Dropbox/shellfish/data/dmr/P90")
}
\arguments{
\item{year}{numeric or string year to read}

\item{path}{local path to files}
}
\value{
sf of Points
}
\description{
Read locally stored P90 data
}
