% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{to_POSIXct}
\alias{to_POSIXct}
\title{Convert ESRI datetime (ms since 1970-01-01) to POSIXct}
\usage{
to_POSIXct(x, origin = "1970-01-01 00:00.00 UTC", fact = 1000)
}
\arguments{
\item{x}{the input numeric ESRI timestamp}

\item{origin}{character - epoch origin as string}

\item{fact}{numeric scaling factor to convert from ms to s}
}
\description{
Convert ESRI datetime (ms since 1970-01-01) to POSIXct
}
