% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uri.R
\name{uri_nssp}
\alias{uri_nssp}
\title{Get the URI for past nssp by year}
\usage{
uri_nssp(year = "2016", base_uri = file.path(uri_base(),
  "DMR_Public_Health_Historical_Bacteria_Data/MapServer"))
}
\arguments{
\item{year}{string or numerc year}

\item{base_uri}{the base MapServer uri}
}
\value{
the uri as string
}
\description{
Get the URI for past nssp by year
}
