% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fetch_current}
\alias{fetch_current}
\title{Fetch the current NSSP closures as geojson}
\usage{
fetch_current(uri = uri_current(),
  dst_path = "/Users/ben/Dropbox/shellfish/data/dmr/current")
}
\arguments{
\item{uri}{string the uri to fetch}

\item{dst_path}{string, the path to write to}
}
\description{
Fetch the current NSSP closures as geojson
}
